changed internal stride types to size_t to prevent issues like (#1419)

+++ libheif/api/libheif/heif_experimental.h
@@ -257,102 +257,102 @@ enum heif_channel_datatype heif_image_get_datatype(const struct heif_image* img,
 LIBHEIF_API
 const uint16_t* heif_image_get_channel_uint16_readonly(const struct heif_image*,
                                                        enum heif_channel channel,
-                                                       uint32_t* out_stride);
+                                                       size_t* out_stride);
 
 LIBHEIF_API
 const uint32_t* heif_image_get_channel_uint32_readonly(const struct heif_image*,
                                                        enum heif_channel channel,
-                                                       uint32_t* out_stride);
+                                                       size_t* out_stride);
 
 LIBHEIF_API
 const uint64_t* heif_image_get_channel_uint64_readonly(const struct heif_image*,
                                                        enum heif_channel channel,
-                                                       uint32_t* out_stride);
+                                                       size_t* out_stride);
 
 LIBHEIF_API
 const int16_t* heif_image_get_channel_int16_readonly(const struct heif_image*,
                                                      enum heif_channel channel,
-                                                     uint32_t* out_stride);
+                                                     size_t* out_stride);
 
 LIBHEIF_API
 const int32_t* heif_image_get_channel_int32_readonly(const struct heif_image*,
                                                      enum heif_channel channel,
-                                                     uint32_t* out_stride);
+                                                     size_t* out_stride);
 
 LIBHEIF_API
 const int64_t* heif_image_get_channel_int64_readonly(const struct heif_image*,
                                                      enum heif_channel channel,
-                                                     uint32_t* out_stride);
+                                                     size_t* out_stride);
 
 LIBHEIF_API
 const float* heif_image_get_channel_float32_readonly(const struct heif_image*,
                                                      enum heif_channel channel,
-                                                     uint32_t* out_stride);
+                                                     size_t* out_stride);
 
 LIBHEIF_API
 const double* heif_image_get_channel_float64_readonly(const struct heif_image*,
                                                       enum heif_channel channel,
-                                                      uint32_t* out_stride);
+                                                      size_t* out_stride);
 
 LIBHEIF_API
 const struct heif_complex32* heif_image_get_channel_complex32_readonly(const struct heif_image*,
                                                                        enum heif_channel channel,
-                                                                       uint32_t* out_stride);
+                                                                       size_t* out_stride);
 
 LIBHEIF_API
 const struct heif_complex64* heif_image_get_channel_complex64_readonly(const struct heif_image*,
                                                                        enum heif_channel channel,
-                                                                       uint32_t* out_stride);
+                                                                       size_t* out_stride);
 
 LIBHEIF_API
 uint16_t* heif_image_get_channel_uint16(struct heif_image*,
                                         enum heif_channel channel,
-                                        uint32_t* out_stride);
+                                        size_t* out_stride);
 
 LIBHEIF_API
 uint32_t* heif_image_get_channel_uint32(struct heif_image*,
                                         enum heif_channel channel,
-                                        uint32_t* out_stride);
+                                        size_t* out_stride);
 
 LIBHEIF_API
 uint64_t* heif_image_get_channel_uint64(struct heif_image*,
                                         enum heif_channel channel,
-                                        uint32_t* out_stride);
+                                        size_t* out_stride);
 
 LIBHEIF_API
 int16_t* heif_image_get_channel_int16(struct heif_image*,
                                       enum heif_channel channel,
-                                      uint32_t* out_stride);
+                                      size_t* out_stride);
 
 LIBHEIF_API
 int32_t* heif_image_get_channel_int32(struct heif_image*,
                                       enum heif_channel channel,
-                                      uint32_t* out_stride);
+                                      size_t* out_stride);
 
 LIBHEIF_API
 int64_t* heif_image_get_channel_int64(struct heif_image*,
                                       enum heif_channel channel,
-                                      uint32_t* out_stride);
+                                      size_t* out_stride);
 
 LIBHEIF_API
 float* heif_image_get_channel_float32(struct heif_image*,
                                       enum heif_channel channel,
-                                      uint32_t* out_stride);
+                                      size_t* out_stride);
 
 LIBHEIF_API
 double* heif_image_get_channel_float64(struct heif_image*,
                                        enum heif_channel channel,
-                                       uint32_t* out_stride);
+                                       size_t* out_stride);
 
 LIBHEIF_API
 struct heif_complex32* heif_image_get_channel_complex32(struct heif_image*,
                                                         enum heif_channel channel,
-                                                        uint32_t* out_stride);
+                                                        size_t* out_stride);
 
 LIBHEIF_API
 struct heif_complex64* heif_image_get_channel_complex64(struct heif_image*,
                                                         enum heif_channel channel,
-                                                        uint32_t* out_stride);
+                                                        size_t* out_stride);
 
 
 // ========================= Timestamps =========================
